requirements.txt
kiteconnect
PyYAML
SQLAlchemy
websocket-client
pandas
pytz
apscheduler
numpy